refactor: move some single-use utils to feature folder
