+++ src/server/features/signature-help/signature-help.ts
@@ -10,13 +10,13 @@ import type {
 	ScssMixin,
 } from "../../parser";
 import type StorageService from "../../storage";
-import { hasInFacts } from "../../utils/facts";
 import { applySassDoc } from "../../utils/sassdoc";
 import {
 	asDollarlessVariable,
 	getTextBeforePosition,
 } from "../../utils/string";
 import { sassBuiltInModules } from "../sass-built-in-modules";
+import { hasInFacts } from "./facts";
 
 // RegExp's
 const reNestedParenthesis = /\(([\w-]+)\(/;
